% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesmaterial.R
\name{buttonkrmesmaterialsrcmatadataUI_left}
\alias{buttonkrmesmaterialsrcmatadataUI_left}
\title{操作区域}
\usage{
buttonkrmesmaterialsrcmatadataUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
