% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesmaterial.R
\name{buttonkrmesmaterialsrcmatadataUI_right}
\alias{buttonkrmesmaterialsrcmatadataUI_right}
\title{预览区域}
\usage{
buttonkrmesmaterialsrcmatadataUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
