% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesprocessplanning.R
\name{buttonkrmesprocessplanningsrcUI_right}
\alias{buttonkrmesprocessplanningsrcUI_right}
\title{预览区域}
\usage{
buttonkrmesprocessplanningsrcUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
