% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesprocessreport.R
\name{buttonkrmesprocessreportsrcUI_left}
\alias{buttonkrmesprocessreportsrcUI_left}
\title{操作区域}
\usage{
buttonkrmesprocessreportsrcUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
