% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesprocess.R
\name{buttonkrmesprocesssrcmatadataUI_right}
\alias{buttonkrmesprocesssrcmatadataUI_right}
\title{预览区域}
\usage{
buttonkrmesprocesssrcmatadataUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
