% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesprocesstransfer.R
\name{buttonkrmesprocesstransfersrcUI_left}
\alias{buttonkrmesprocesstransfersrcUI_left}
\title{操作区域}
\usage{
buttonkrmesprocesstransfersrcUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
