% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesprocesstransfer.R
\name{buttonkrmesprocesstransfersrcUI_right}
\alias{buttonkrmesprocesstransfersrcUI_right}
\title{预览区域}
\usage{
buttonkrmesprocesstransfersrcUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
