% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionfeed.R
\name{buttonkrmesproductionfeedsrcUI_left}
\alias{buttonkrmesproductionfeedsrcUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionfeedsrcUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
