% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionfeed.R
\name{buttonkrmesproductionfeedsrcUI_right}
\alias{buttonkrmesproductionfeedsrcUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionfeedsrcUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
