% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionwarehousin.R
\name{buttonkrmesproductionwarehousinsrcUI_left}
\alias{buttonkrmesproductionwarehousinsrcUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionwarehousinsrcUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
