% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesroute.R
\name{buttonkrmesroutesrcmatadataUI_right}
\alias{buttonkrmesroutesrcmatadataUI_right}
\title{预览区域}
\usage{
buttonkrmesroutesrcmatadataUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
