% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesstock.R
\name{buttonkrmesstocksrcmatadataUI_right}
\alias{buttonkrmesstocksrcmatadataUI_right}
\title{预览区域}
\usage{
buttonkrmesstocksrcmatadataUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
